+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Todos los cambios notables en este proyecto serán documentados en este archivo.
 El formato está basado en [Keep a Changelog](https://keepachangelog.com/es-ES/1.0.0/),
 y este proyecto adhiere a [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2024-03-15
+
+### Agregado
+- Al limpiar el calendario y volver a abrirlo mostrar la fecha actual
+- Corrección de posicion en modo dropdown
+
 ## [1.0.0] - 2024-03-15
 
 ### Agregado